+++ src/private/Get-DefaultBuildFile.ps1
@@ -6,12 +6,12 @@ function Get-DefaultBuildFile {
         [boolean] $UseDefaultIfNoneExist = $true
     )
 
-    if (test-path $psake.config_default.buildFileName -pathType Leaf) {
-        Write-Output $psake.config_default.buildFileName
-    } elseif (test-path $psake.config_default.legacyBuildFileName -pathType Leaf) {
+    if (Test-Path $psake.ConfigDefault.buildFileName -PathType Leaf) {
+        Write-Output $psake.ConfigDefault.buildFileName
+    } elseif (Test-Path $psake.ConfigDefault.legacyBuildFileName -PathType Leaf) {
         Write-Warning "The default configuration file of default.ps1 is deprecated.  Please use psakefile.ps1"
-        Write-Output $psake.config_default.legacyBuildFileName
+        Write-Output $psake.ConfigDefault.legacyBuildFileName
     } elseif ($UseDefaultIfNoneExist) {
-        Write-Output $psake.config_default.buildFileName
+        Write-Output $psake.ConfigDefault.buildFileName
     }
 }